Split panel switcher out
